pyproject.toml
[project]
requires-python = ">= 3.10"
dependencies = [
    "reboot-resemble>=0.10.0",
]

[tool.rye]
dev-dependencies = [
  "reboot-resemble>=0.10.0",
  "mypy==1.2.0",
  "pytest>=7.4.2",
  "types-protobuf>=4.24.0.20240129",
]

# This project only uses `rye` to provide `python` and its dependencies, so
# these settings remove the need to name or version the project.
virtual = true
managed = true
